Improved network code + localized network object creation
