--- Assets/Scripts/Config.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1eb4f9755824c994588f211426e3709a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
