+++ Assets/Scripts/Network/INetworkSendable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 870025e5a71278c4ba66d831877fd7c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
