+++ Assets/Scripts/NetworkInitializer.cs
@@ -1,13 +1,10 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using DarkRift;
+﻿using DarkRift;
 using DarkRift.Client.Unity;
 using DarkRift.Server.Unity;
-using System.Net;
+using Network;
+using UnityEngine;
 
 public class NetworkInitializer : MonoBehaviour {
-    
     [Tooltip("NetworkConfig to read configurations from")]
     public NetworkConfig NetworkConfig;
 
@@ -19,7 +16,6 @@ public class NetworkInitializer : MonoBehaviour {
 
     [Tooltip("Event to trigger after server is initialized")]
     public GameEvent Event;
-
 
     void Start() {
         if (Client == null || NetworkConfig == null || Server == null) {
@@ -34,5 +30,4 @@ public class NetworkInitializer : MonoBehaviour {
 
         Client.Connect(NetworkConfig.ip, NetworkConfig.port, IPVersion.IPv4);
     }
-
 }
